fix: allow pip to install globally
When using pip install, it really wants us to use virtual
python environments, and will fail if we try to pip install the Azure
CLI globally. Since we have a very constrained use case for the runner,
we are OK with installing globally, as there's nothing else using the
Azure CLI to clobber.